+++ CHANGELOG.md
@@ -3180,3 +3180,4 @@
 - [2026-07-27 12:55] refactor(storage): decouple repository interface from concrete storage service
 - [2026-07-27 12:56] refactor(ui): decouple repository interface from concrete ui service
 - [2026-07-27 13:00] chore(storage): update dependencies and cleanup unused imports in storage
+- [2026-07-27 13:22] chore(ui): update dependencies and cleanup unused imports in ui